refactoring : pages main components in index files with default exports.
