--- src/features/users/index.ts
@@ -1,2 +1,0 @@
-export { default as UserCardList } from './ui/UserCardList';
-export { default as UsersPagination } from './ui/UsersPagination';
