--- src/features/users/model/useUsersQueryParams.ts
@@ -1,12 +1,0 @@
-import { useAppSelector } from '@shared/redux';
-import { selectCurrentPage, selectUsersQueryCount } from './usersSlice';
-import { UsersQueryParams } from './types';
-
-const useUsersQueryParams = (): UsersQueryParams => {
-  const usersQueryCount = useAppSelector(selectUsersQueryCount);
-  const currentPage = useAppSelector(selectCurrentPage);
-
-  return { usersQueryCount, currentPage };
-};
-
-export default useUsersQueryParams;
