--- src/pages/login-page/index.ts
@@ -1,1 +1,0 @@
-export { default as LoginPage } from './ui/LoginPage';
