--- src/pages/not-found/index.ts
@@ -1,1 +1,0 @@
-export { default as NotFoundPage } from './ui/NotFoundPage';
